+++ src/config/routes.ts
@@ -28,12 +28,11 @@ export const relativeRoutes = {
     index: () => '',
     newChannel: () => 'channel/new',
     editChannel: () => 'channel',
-    videos: () => '/videos',
+    videos: () => 'videos',
     uploads: () => 'uploads',
     signIn: () => 'signIn',
     newMembership: () => 'membership/new',
     selectMembership: () => 'memberships',
-    selectChannel: () => 'channels',
   },
   playground: {
     index: () => '',
