--- src/shared/components/AccountBar/index.ts
@@ -1,2 +1,0 @@
-import AccountBar from './AccountBar'
-export default AccountBar
