+++ src/shared/components/Spinner/Spinner.tsx
@@ -1,12 +1,14 @@
 import React from 'react'
 import { SpinnerWrapper } from './Spinner.style'
 
+export type SpinnerSize = 'large' | 'medium' | 'small'
+
 export type SpinnerProps = {
-  size?: number
+  size?: SpinnerSize
   className?: string
 }
 
-const Spinner: React.FC<SpinnerProps> = ({ className, size }) => {
+const Spinner: React.FC<SpinnerProps> = ({ className, size = 'medium' }) => {
   return <SpinnerWrapper className={className} size={size} />
 }
