--- src/shared/components/StudioHeader/index.ts
@@ -1,3 +1,0 @@
-import StudioHeader from './StudioHeader'
-
-export default StudioHeader
