+++ src/shared/components/index.ts
@@ -36,7 +36,4 @@ export { default as ActionBar, ActionBarTransaction } from './ActionBar'
 export { default as MultiFileSelect } from './MultiFileSelect'
 export { default as FileSelect } from './FileSelect'
 export { default as FileStep } from './FileStep'
-export { default as StudioCard } from './StudioCard'
-export { default as StudioHeader } from './StudioHeader'
 export { default as Spinner } from './Spinner'
-export { default as AccountBar } from './AccountBar'
