+++ src/views/studio/CreateMemberView/CreateMemberView.tsx
@@ -1,14 +1,14 @@
 import { ActionDialog } from '@/components/Dialogs'
 import { absoluteRoutes } from '@/config/routes'
 import { useActiveUser } from '@/hooks'
-import { Spinner, StudioHeader, Text, TextField } from '@/shared/components'
+import { Spinner, Text, TextField } from '@/shared/components'
 import TextArea from '@/shared/components/TextArea'
 import { textFieldValidation, urlValidation } from '@/utils/formValidationOptions'
 import React, { useEffect, useState } from 'react'
 import { useForm } from 'react-hook-form'
 import { useNavigate } from 'react-router'
 import { createFakeMembership } from '../SignInView/fakeUtils'
-import { Form, StyledButton, Wrapper, StyledText } from './CreateMemberView.style'
+import { Form, StyledButton, Wrapper, StyledText, Header, Hero, SubTitle } from './CreateMemberView.style'
 
 type Inputs = {
   handle: string
@@ -56,10 +56,12 @@ const CreateMemberView = () => {
 
   return (
     <Wrapper>
-      <StudioHeader
-        title="Add Member Details"
-        subtitle="Customize your membership details to easily spot the member account you would like to use when signing in."
-      />
+      <Header>
+        <Hero variant="hero">Add Member Details</Hero>
+        <SubTitle variant="body2">
+          Customize your membership details to easily spot the member account you would like to use when signing in.
+        </SubTitle>
+      </Header>
       <Form onSubmit={onSubmit}>
         <TextField
           name="avatarUri"
