+++ src/views/studio/SelectMembershipView/StudioCard/index.ts
